+++ Assets/Countly/Plugins/CountlySDK/Helpers/RequestCountlyHelper.cs
@@ -19,13 +19,11 @@ namespace Plugins.CountlySDK.Helpers
     public class RequestCountlyHelper
     {
         private bool _isQueueBeingProcess;
-
         private readonly CountlyLogHelper Log;
         private readonly CountlyUtils _countlyUtils;
         private readonly CountlyConfiguration _config;
         private readonly RequestBuilder _requestBuilder;
         internal readonly RequestRepository _requestRepo;
-
         private readonly MonoBehaviour _monoBehaviour;
         private readonly int countlyRequestRemoveLimit = 100;
 
@@ -46,7 +44,7 @@ namespace Plugins.CountlySDK.Helpers
         /// <param name="request">The Countly request model to be added to the queue.</param>
         internal void AddRequestToQueue(CountlyRequestModel request)
         {
-            Log.Verbose("[RequestCountlyHelper] AddRequestToQueue: " + request.ToString());
+            Log.Verbose($"[RequestCountlyHelper] AddRequestToQueue, Request: [{request.ToString()}]");
 
             if (_config.EnableTestMode) {
                 return;
@@ -75,22 +73,38 @@ namespace Plugins.CountlySDK.Helpers
         internal async Task ProcessQueue()
         {
             if (_isQueueBeingProcess) {
+                Log.Verbose("[RequestCountlyHelper] ProcessQueue, queue is being processed.");
                 return;
             }
 
             _isQueueBeingProcess = true;
-            CountlyRequestModel[] requests = _requestRepo.Models.ToArray();
+            CountlyRequestModel[] requests = null;
 
-            Log.Verbose("[RequestCountlyHelper] Process queue, requests: " + requests.Length);
+            try {
+                requests = _requestRepo.Models.ToArray();
+                Log.Verbose("[RequestCountlyHelper] ProcessQueue, request count: " + requests.Length);
+            } catch (Exception ex) {
+                Log.Warning($"[RequestCountlyHelper] ProcessQueue, exception occurred while converting Models to array. Exception: {ex}");
+                _isQueueBeingProcess = false;
+                return;
+            }
+
+            // make sure that flag is corrected and exit
+            if (requests.Length == 0) {
+                Log.Verbose("[RequestCountlyHelper] ProcessQueue, Request Queue is empty. Returning.");
+                _isQueueBeingProcess = false;
+                return;
+            }
 
             foreach (CountlyRequestModel reqModel in requests) {
-                //add the remaining request count in RequestData
+                // Add the remaining request count in RequestData
                 reqModel.RequestData += "&rr=" + (requests.Length - 1);
 
                 CountlyResponse response = await ProcessRequest(reqModel);
 
                 if (!response.IsSuccess) {
-                    Log.Verbose("[RequestCountlyHelper] ProcessQueue: Request fail, " + response.ToString());
+                    Log.Verbose($"[RequestCountlyHelper] ProcessQueue: Request failed. Response: [{response.ToString()}]");
+                    _isQueueBeingProcess = false;
                     break;
                 }
 
@@ -98,6 +112,7 @@ namespace Plugins.CountlySDK.Helpers
             }
 
             _isQueueBeingProcess = false;
+            Log.Verbose("[RequestCountlyHelper] ProcessQueue, _isQueueBeingProcess is false");
         }
 
         /// <summary>
@@ -105,22 +120,25 @@ namespace Plugins.CountlySDK.Helpers
         /// </summary>
         private async Task<CountlyResponse> ProcessRequest(CountlyRequestModel model)
         {
-            Log.Verbose("[RequestCountlyHelper] Process request, request: " + model);
+            Log.Verbose($"[RequestCountlyHelper] Process request, request: [{model}]");
             bool shouldPost = _config.IsForcedHttpPostEnabled() || model.RequestData.Length > 2000;
 
 #if UNITY_WEBGL
-            // There is not HTTP GET for webGL. We always do a HTTP POST
+            // There is not HTTP GET for WebGL. We always do a HTTP POST
+            Log.Debug($"[RequestCountlyHelper] ProcessRequest, Calling StartProcessRequestRoutine for WebGL");
             return await StartProcessRequestRoutine(_countlyUtils.ServerInputUrl, model.RequestData);
 #else
             if (shouldPost) {
+                Log.Debug($"[RequestCountlyHelper] ProcessRequest, Using HTTP POST");
                 return await Task.Run(() => PostAsync(_countlyUtils.ServerInputUrl, model.RequestData));
             }
+            Log.Debug($"[RequestCountlyHelper] ProcessRequest, Using HTTP GET");
             return await Task.Run(() => GetAsync(_countlyUtils.ServerInputUrl, model.RequestData));
 #endif
         }
 
         /// <summary>
-        ///  An internal function to add a request to request queue.
+        /// An internal function to add a request to request queue.
         /// </summary>
         internal void AddToRequestQueue(Dictionary<string, object> queryParams)
         {
@@ -138,7 +156,7 @@ namespace Plugins.CountlySDK.Helpers
                     string hex = _countlyUtils.GetStringFromBytes(bytes);
 
                     query += "&checksum256=" + hex;
-                    Log.Debug("BuildGetRequest: query = " + query);
+                    Log.Debug($"[RequestCountlyHelper] AddChecksum, Checksum added query = [{query}]");
                 }
             }
 
@@ -153,7 +171,7 @@ namespace Plugins.CountlySDK.Helpers
         /// <returns></returns>
         internal async Task<CountlyResponse> GetAsync(string uri, string data)
         {
-            Log.Verbose("[RequestCountlyHelper] GetAsync request: " + uri + " params: " + data);
+            Log.Verbose($"[RequestCountlyHelper] GetAsync, calling with request: [{uri}], params: [{data}]");
 
             CountlyResponse countlyResponse = new CountlyResponse();
             string query = AddChecksum(data);
@@ -171,7 +189,6 @@ namespace Plugins.CountlySDK.Helpers
                         countlyResponse.StatusCode = code;
                         countlyResponse.IsSuccess = IsSuccess(countlyResponse);
                     }
-
                 }
             } catch (WebException ex) {
                 countlyResponse.ErrorMessage = ex.Message;
@@ -188,8 +205,7 @@ namespace Plugins.CountlySDK.Helpers
                 }
             }
 
-            Log.Verbose("[RequestCountlyHelper] GetAsync request: " + url + " params: " + query + " response: " + countlyResponse.ToString());
-
+            Log.Verbose($"[RequestCountlyHelper] GetAsync, Request URL: [{url}], Params: [{query}], Response: [{countlyResponse.ToString()}]");
             return countlyResponse;
         }
 
@@ -242,8 +258,7 @@ namespace Plugins.CountlySDK.Helpers
                 }
             }
 
-            Log.Verbose("[RequestCountlyHelper] PostAsync request: " + uri + " body: " + data + " response: " + countlyResponse.ToString());
-
+            Log.Verbose($"[RequestCountlyHelper] PostAsync, Request URL: [{uri}], Body: [{data}], Response: [{countlyResponse.ToString()}]");
             return countlyResponse;
         }
 
@@ -255,11 +270,26 @@ namespace Plugins.CountlySDK.Helpers
         /// <returns></returns>
         private Task<CountlyResponse> StartProcessRequestRoutine(string uri, string data)
         {
+            Log.Debug($"[RequestCountlyHelper] StartProcessRequestRoutine, Start");
             TaskCompletionSource<CountlyResponse> tcs = new TaskCompletionSource<CountlyResponse>();
 
-            _monoBehaviour.StartCoroutine(ProcessRequestCoroutine(uri, data, (response) => {
-                tcs.SetResult(response);
-            }));
+            try {
+                // 'IsObjectMonoBehaviour can only be called from the main thread'
+                // That's why we have to move it to main thread.
+                CountlyMainThreadHandler.Instance.RunOnMainThread(() => {
+                    try {
+                        _monoBehaviour.StartCoroutine(ProcessRequestCoroutine(uri, data, (response) => {
+                            tcs.SetResult(response);
+                        }));
+                    } catch (Exception ex) {
+                        Log.Error($"[RequestCountlyHelper] StartProcessRequestRoutine, Exception occurred while starting coroutine. Exception: [{ex}]");
+                        tcs.SetException(ex);
+                    }
+                });
+            } catch (Exception ex) {
+                Log.Error($"[RequestCountlyHelper] StartProcessRequestRoutine, Exception occurred while queueing to main thread. Exception: [{ex}]");
+                tcs.SetException(ex);
+            }
 
             return tcs.Task;
         }
@@ -273,17 +303,15 @@ namespace Plugins.CountlySDK.Helpers
         /// <returns></returns>
         private IEnumerator ProcessRequestCoroutine(string uri, string data, Action<CountlyResponse> callback)
         {
+            Log.Debug($"[RequestCountlyHelper] ProcessRequestCoroutine, Start");
             CountlyResponse countlyResponse = new CountlyResponse();
 
             string query = AddChecksum(data);
             string url = uri + query;
 
             using (UnityWebRequest webRequest = UnityWebRequest.Put(url, data)) {
-
                 webRequest.method = UnityWebRequest.kHttpVerbPOST;
-
                 yield return webRequest.SendWebRequest();
-
                 string[] pages = url.Split('?');
                 int page = pages.Length - 1;
                 int code = (int)webRequest.responseCode;
@@ -292,13 +320,17 @@ namespace Plugins.CountlySDK.Helpers
                 countlyResponse.ErrorMessage = webRequest.error;
                 countlyResponse.StatusCode = code;
                 countlyResponse.IsSuccess = IsSuccess(countlyResponse);
-
-                Log.Debug($"[RequestCountlyHelper] ProcessRequestCoroutine request url: [{uri}] body: [{pages[page]}] response: [{countlyResponse}]");
+                Log.Debug($"[RequestCountlyHelper] ProcessRequestCoroutine, Request URL: [{uri}], Body: [{pages[page]}], Response: [{countlyResponse}]");
             }
-
+            Log.Debug($"[RequestCountlyHelper] ProcessRequestCoroutine, Process completed invoking callback.");
             callback?.Invoke(countlyResponse);
         }
 
+        /// <summary>
+        /// Determines if a CountlyResponse indicates a successful request based on the HTTP status code and the presence of a "result" key in the JSON response data.
+        /// </summary>
+        /// <param name="countlyResponse">The CountlyResponse object to evaluate.</param>
+        /// <returns>True if the status code is between 200 and 299 and the response data contains a "result" key; otherwise, false.</returns>
         private bool IsSuccess(CountlyResponse countlyResponse)
         {
             if (countlyResponse.StatusCode >= 200 && countlyResponse.StatusCode < 300) {
@@ -308,12 +340,13 @@ namespace Plugins.CountlySDK.Helpers
                     if (json.ContainsKey("result")) {
                         return true;
                     }
-                } catch(JsonException) {
-                    Log.Debug("[RequestCountlyHelper] IsSuccess : Returned request is not a JSON object");
+                } catch (JsonException ex) {
+                    Log.Debug($"[RequestCountlyHelper] IsSuccess : Returned request is not a JSON object. Exception: [{ex}]");
                     return false;
-                } 
+                }
+            } else {
+                Log.Debug($"[RequestCountlyHelper] IsSuccess, Status Code: [{countlyResponse.StatusCode}] is not in between 200-299. Returning false.");
             }
-
             return false;
         }
     }