+++ Assets/Countly/Plugins/CountlySDK/Services/SessionCountlyService.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using System.Timers;
@@ -6,6 +7,7 @@ using Newtonsoft.Json;
 using Plugins.CountlySDK.Enums;
 using Plugins.CountlySDK.Helpers;
 using Plugins.CountlySDK.Models;
+using UnityEngine;
 
 namespace Plugins.CountlySDK.Services
 {
@@ -22,11 +24,11 @@ namespace Plugins.CountlySDK.Services
         private readonly LocationService _locationService;
         private readonly EventCountlyService _eventService;
         internal readonly RequestCountlyHelper _requestCountlyHelper;
-
+        private readonly MonoBehaviour _monoBehaviour;
         bool isInternalTimerStopped;
 
         internal SessionCountlyService(CountlyConfiguration configuration, CountlyLogHelper logHelper, EventCountlyService eventService,
-            RequestCountlyHelper requestCountlyHelper, LocationService locationService, ConsentCountlyService consentService) : base(configuration, logHelper, consentService)
+            RequestCountlyHelper requestCountlyHelper, LocationService locationService, ConsentCountlyService consentService, MonoBehaviour monoBehaviour) : base(configuration, logHelper, consentService)
         {
             Log.Debug("[SessionCountlyService] Initializing.");
             if (configuration.IsAutomaticSessionTrackingDisabled) {
@@ -36,6 +38,7 @@ namespace Plugins.CountlySDK.Services
             _eventService = eventService;
             _locationService = locationService;
             _requestCountlyHelper = requestCountlyHelper;
+            _monoBehaviour = monoBehaviour;
 
             if (_configuration.IsAutomaticSessionTrackingDisabled) {
                 Log.Verbose("[Countly][CountlyConfiguration] Automatic session tracking disabled!");
@@ -72,10 +75,38 @@ namespace Plugins.CountlySDK.Services
         /// </summary>
         private void InitSessionTimer()
         {
+#if UNITY_WEBGL
+            _monoBehaviour.StartCoroutine(SessionTimerCoroutine());
+#else
             _sessionTimer = new Timer { Interval = _configuration.GetUpdateSessionTimerDelay() * 1000 };
             _sessionTimer.Elapsed += SessionTimerOnElapsedAsync;
             _sessionTimer.AutoReset = true;
             _sessionTimer.Start();
+#endif
+        }
+
+        private void SendRequestsAndExtendSession()
+        {
+            Countly.Instance.UserProfile.Save();
+            _eventService.AddEventsToRequestQueue();
+            _ = _requestCountlyHelper.ProcessQueue();
+
+            if (!_configuration.IsAutomaticSessionTrackingDisabled) {
+                _ = ExtendSessionAsync();
+            }
+        }
+
+        private IEnumerator SessionTimerCoroutine()
+        {
+            Log.Debug("[SessionCountlyService] SessionTimerCoroutine, Start");
+
+            if (isInternalTimerStopped) {
+                yield break;
+            }
+
+            yield return new WaitForSeconds(_configuration.GetUpdateSessionTimerDelay());
+            SendRequestsAndExtendSession();
+            Log.Debug("[SessionCountlyService] SessionTimerCoroutine, Coroutine completed.");
         }
 
         /// <summary>
@@ -86,8 +117,10 @@ namespace Plugins.CountlySDK.Services
         {
             isInternalTimerStopped = true;
 
-            if (_sessionTimer != null)
-            {
+            #if UNITY_WEBGL
+            _monoBehaviour.StopCoroutine(SessionTimerCoroutine());
+            #else
+            if (_sessionTimer != null) {
                 // Unsubscribe from the Elapsed event
                 _sessionTimer.Elapsed -= SessionTimerOnElapsedAsync;
 
@@ -95,6 +128,7 @@ namespace Plugins.CountlySDK.Services
                 _sessionTimer.Stop();
                 _sessionTimer.Dispose();
             }
+            #endif
         }
 
         /// <summary>
@@ -111,15 +145,7 @@ namespace Plugins.CountlySDK.Services
                 }
 
                 Log.Debug("[SessionCountlyService] SessionTimerOnElapsedAsync");
-
-                _eventService.AddEventsToRequestQueue();
-                Countly.Instance.UserProfile.Save();
-                
-                _ = _requestCountlyHelper.ProcessQueue();
-
-                if (!_configuration.IsAutomaticSessionTrackingDisabled) {
-                    _ = ExtendSessionAsync();
-                }
+                SendRequestsAndExtendSession();
             }
 
             await Task.CompletedTask;
@@ -192,7 +218,7 @@ namespace Plugins.CountlySDK.Services
                 Log.Warning("[SessionCountlyService] EndSessionAsync: The session isn't started yet!");
                 return;
             }
-            
+
             IsSessionInitiated = false;
             _eventService.AddEventsToRequestQueue();
             Countly.Instance.UserProfile.Save();
@@ -202,11 +228,9 @@ namespace Plugins.CountlySDK.Services
                     {"session_duration",  Convert.ToInt32((DateTime.Now - _lastSessionRequestTime).TotalSeconds)}
                 };
 
-
             _requestCountlyHelper.AddToRequestQueue(requestParams);
             await _requestCountlyHelper.ProcessQueue();
         }
-
 
         /// <summary>
         /// Extends a session by another session duration provided in configuration. By default session duration is 60 seconds.
@@ -235,6 +259,10 @@ namespace Plugins.CountlySDK.Services
 
             _requestCountlyHelper.AddToRequestQueue(requestParams);
             await _requestCountlyHelper.ProcessQueue();
+            
+            #if UNITY_WEBGL
+            _monoBehaviour.StartCoroutine(SessionTimerCoroutine());
+            #endif
         }
 
         #region override Methods