+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 * Deprecated "Countly.Instance.UserDetails" added "Countly.Instance.UserProfile" as replacement, mentioned above. 
 * Fixed a bug that allowed to make it possible to close non-started views.
 * Fixed issues that allowed to record User Profiles without consent.
+* Fixed a bug that caused requests being stuck in the queue for WebGL build targets.
 
 ## 23.12.1
 * Added 'UnityWebRequest' as the networking handler for WebGL builds.